Added the two datasets from the surf publication (#195)
* Added the two datasets from the surf publication

Borylation dataset and minisci dataset.

* Create test_file

Adding the file to make a commit. It will be subsequently removed in a new commit. I think this will reset the check target branch test.

* Delete data/test_file

removing this file because it is has served its purpose. The check target branch test was reset to see the correct target branch.