Added the two datasets from the surf publication (#195) (#199)
* Added the two datasets from the surf publication (#195)

* Added the two datasets from the surf publication

Borylation dataset and minisci dataset.

* Create test_file

Adding the file to make a commit. It will be subsequently removed in a new commit. I think this will reset the check target branch test.

* Delete data/test_file

removing this file because it is has served its purpose. The check target branch test was reset to see the correct target branch.

* Update submission

* Create test_file.txt

* Delete test_file.txt

* Update badges

* Update badges

* updated the minisci dataset

* Update submission

* Revert "updated the minisci dataset"

This reverts commit b8ca10066ebbbd91ac53b0299c1628da92aa1a80.

* Update badges

* replaced minisci

* Update submission

* Update badges

* Update badges

---------

Co-authored-by: github-actions <github-actions@github.com>
Co-authored-by: Steven Kearnes <skearnes@relaytx.com>